+++ CHANGELOG.md
@@ -61,6 +61,10 @@ Playground will have all `derive_more` features available once
   instead of replacing when no wrapping is possible (no `_variant` placeholder).
   ([#395](https://github.com/JelteF/derive_more/pull/395))
 
+### Added
+
+- Derive for `AsVariant`.
+
 ### Fixed
 
 - Associated types of type parameters not being treated as generics in `Debug`